+++ src/lib/ui/core/components/text_input.py
@@ -4,10 +4,10 @@ from direct.gui.DirectEntry import DirectEntry
 from panda3d.core import DynamicTextFont, NodePath, PandaNode
 
 from lib.app.focus.focusable import Focusable
+from lib.ui.context import UIContext
 from lib.ui.core.alignment import HAlign, VAlign
 from lib.ui.core.colors import UIColors
 from lib.ui.core.constants import UIConstants
-from lib.ui.core.context import UIContext
 from lib.ui.core.layers import UILayer
 from lib.ui.core.util import correctYForTextAlignment, horizontalAlignToTextNodeAlign
 from lib.util.events.event_dispatcher import EventDispatcher
