+++ src/lib/ui/header/clock.py
@@ -3,10 +3,10 @@ import datetime
 from direct.showbase.DirectObject import DirectObject
 from direct.task.Task import Task
 
+from lib.ui.context import UIContext
 from lib.ui.core.alignment import HAlign, VAlign
 from lib.ui.core.components.text import Text
 from lib.ui.core.constants import UIConstants
-from lib.ui.core.context import UIContext
 
 
 class Clock(DirectObject):
