+++ src/lib/ui/panels/components/label.py
@@ -2,10 +2,10 @@ from __future__ import annotations
 
 from panda3d.core import NodePath, PandaNode
 
+from lib.ui.context import UIContext
 from lib.ui.core.alignment import HAlign, VAlign
 from lib.ui.core.components.text import Text
 from lib.ui.core.constants import UIConstants
-from lib.ui.core.context import UIContext
 
 
 class ComponentLabel:
