+++ src/lib/ui/panels/core/panel_content.py
@@ -2,11 +2,11 @@ from abc import ABC, abstractmethod
 from typing import List, TypeVar
 
 from lib.app.state import AppState
+from lib.ui.context import UIContext
 from lib.ui.core.alignment import HAlign, VAlign
 from lib.ui.core.colors import UIColors
 from lib.ui.core.components.background_card import BackgroundCard
 from lib.ui.core.constants import UIConstants
-from lib.ui.core.context import UIContext
 from lib.ui.core.layers import UILayer
 from lib.ui.events import UIEvents
 from lib.ui.panels.core.panel_component import PanelComponent
