+++ src/lib/ui/panels/settings/ui_scale.py
@@ -3,11 +3,11 @@ from __future__ import annotations
 from panda3d.core import NodePath, PandaNode
 
 from lib.app.state import AppState
+from lib.ui.context import UIContext
 from lib.ui.core.alignment import HAlign, VAlign
 from lib.ui.core.components.text import Text
 from lib.ui.core.components.text_input import TextInput
 from lib.ui.core.constants import UIConstants
-from lib.ui.core.context import UIContext
 from lib.ui.panels.components.label import ComponentLabel
 from lib.ui.panels.core.panel_component import PanelComponent
