+++ src/lib/ui/ui.py
@@ -1,7 +1,7 @@
 from lib.app.context import AppContext
 from lib.app.events import AppEvents
 from lib.app.state import AppState
-from lib.ui.core.context import UIContext
+from lib.ui.context import UIContext
 from lib.ui.footer.footer import Footer
 from lib.ui.header.header import Header
 from lib.ui.panels.panel_module import PanelModule
